requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
easyocr==1.7.2
exceptiongroup==1.2.2
faiss-cpu==1.11.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.54.1
fsspec==2024.10.0
fuzzywuzzy==0.18.0
gradio==5.5.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.30.2
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
omegaconf==2.3.0
openai==1.76.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==4.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2024.9.20
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.51.3
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
Werkzeug==3.1.3
